+++ src/org/jgroups/protocols/PingHeader.java
@@ -1,4 +1,4 @@
-// $Id: PingHeader.java,v 1.7 2004/10/08 13:15:46 belaban Exp $
+// $Id: PingHeader.java,v 1.8 2004/10/08 13:57:32 belaban Exp $
 
 package org.jgroups.protocols;
 
@@ -9,16 +9,16 @@ import java.io.*;
 
 
 public class PingHeader extends Header implements Streamable {
-    public static final short GET_MBRS_REQ=1;   // arg = null
-    public static final short GET_MBRS_RSP=2;   // arg = PingRsp(local_addr, coord_addr)
+    public static final byte GET_MBRS_REQ=1;   // arg = null
+    public static final byte GET_MBRS_RSP=2;   // arg = PingRsp(local_addr, coord_addr)
 
-    public short type=0;
+    public byte type=0;
     public PingRsp arg=null;
 
     public PingHeader() {
     } // for externalization
 
-    public PingHeader(short type, PingRsp arg) {
+    public PingHeader(byte type, PingRsp arg) {
         this.type=type;
         this.arg=arg;
     }
@@ -31,7 +31,7 @@ public class PingHeader extends Header implements Streamable {
         return "[PING: type=" + type2Str(type) + ", arg=" + arg + ']';
     }
 
-    String type2Str(int t) {
+    String type2Str(byte t) {
         switch(t) {
             case GET_MBRS_REQ:
                 return "GET_MBRS_REQ";
@@ -44,18 +44,18 @@ public class PingHeader extends Header implements Streamable {
 
 
     public void writeExternal(ObjectOutput out) throws IOException {
-        out.writeShort(type);
+        out.writeByte(type);
         out.writeObject(arg);
     }
 
 
     public void readExternal(ObjectInput in) throws IOException, ClassNotFoundException {
-        type=in.readShort();
+        type=in.readByte();
         arg=(PingRsp)in.readObject();
     }
 
     public void writeTo(DataOutputStream outstream) throws IOException {
-        outstream.writeShort(type);
+        outstream.writeByte(type);
         if(arg != null) {
             outstream.write(1);
             arg.writeTo(outstream);
@@ -66,7 +66,7 @@ public class PingHeader extends Header implements Streamable {
     }
 
     public void readFrom(DataInputStream instream) throws IOException, IllegalAccessException, InstantiationException {
-        type=instream.readShort();
+        type=instream.readByte();
         int b=instream.read();
         if(b == 1) {
             arg=new PingRsp();
